+++ familyhelper-project-impl/src/main/java/com/dwarfeng/familyhelper/project/impl/cache/TaskTypeIndicatorCacheImpl.java
@@ -1,0 +1,99 @@
+package com.dwarfeng.familyhelper.project.impl.cache;
+
+import com.dwarfeng.familyhelper.project.sdk.bean.entity.FastJsonTaskTypeIndicator;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.TaskTypeIndicator;
+import com.dwarfeng.familyhelper.project.stack.cache.TaskTypeIndicatorCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.CacheException;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class TaskTypeIndicatorCacheImpl implements TaskTypeIndicatorCache {
+
+    private final RedisBatchBaseCache<StringIdKey, TaskTypeIndicator, FastJsonTaskTypeIndicator>
+            taskTypeIndicatorBatchBaseDelegate;
+
+    public TaskTypeIndicatorCacheImpl(
+            RedisBatchBaseCache<StringIdKey, TaskTypeIndicator, FastJsonTaskTypeIndicator>
+                    taskTypeIndicatorBatchBaseDelegate
+    ) {
+        this.taskTypeIndicatorBatchBaseDelegate = taskTypeIndicatorBatchBaseDelegate;
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean exists(StringIdKey key) throws CacheException {
+        return taskTypeIndicatorBatchBaseDelegate.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public TaskTypeIndicator get(StringIdKey key) throws CacheException {
+        return taskTypeIndicatorBatchBaseDelegate.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void push(TaskTypeIndicator value, long timeout) throws CacheException {
+        taskTypeIndicatorBatchBaseDelegate.push(value, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void delete(StringIdKey key) throws CacheException {
+        taskTypeIndicatorBatchBaseDelegate.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void clear() throws CacheException {
+        taskTypeIndicatorBatchBaseDelegate.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return taskTypeIndicatorBatchBaseDelegate.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return taskTypeIndicatorBatchBaseDelegate.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<TaskTypeIndicator> batchGet(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        return taskTypeIndicatorBatchBaseDelegate.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchPush(@SkipRecord List<TaskTypeIndicator> entities, long timeout) throws CacheException {
+        taskTypeIndicatorBatchBaseDelegate.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws CacheException {
+        taskTypeIndicatorBatchBaseDelegate.batchDelete(keys);
+    }
+}