+++ familyhelper-project-impl/src/main/java/com/dwarfeng/familyhelper/project/impl/dao/TaskTypeIndicatorDaoImpl.java
@@ -1,0 +1,136 @@
+package com.dwarfeng.familyhelper.project.impl.dao;
+
+import com.dwarfeng.familyhelper.project.impl.bean.entity.HibernateTaskTypeIndicator;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.TaskTypeIndicator;
+import com.dwarfeng.familyhelper.project.stack.dao.TaskTypeIndicatorDao;
+import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
+import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateStringIdKey;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.SkipRecord;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import com.dwarfeng.subgrade.stack.exception.DaoException;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class TaskTypeIndicatorDaoImpl implements TaskTypeIndicatorDao {
+
+    private final HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, TaskTypeIndicator,
+            HibernateTaskTypeIndicator> batchBaseDao;
+    private final HibernateEntireLookupDao<TaskTypeIndicator, HibernateTaskTypeIndicator> entireLookupDao;
+
+    public TaskTypeIndicatorDaoImpl(
+            HibernateBatchBaseDao<StringIdKey, HibernateStringIdKey, TaskTypeIndicator,
+                    HibernateTaskTypeIndicator> batchBaseDao,
+            HibernateEntireLookupDao<TaskTypeIndicator, HibernateTaskTypeIndicator> entireLookupDao
+    ) {
+        this.batchBaseDao = batchBaseDao;
+        this.entireLookupDao = entireLookupDao;
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public StringIdKey insert(TaskTypeIndicator element) throws DaoException {
+        return batchBaseDao.insert(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void update(TaskTypeIndicator element) throws DaoException {
+        batchBaseDao.update(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void delete(StringIdKey key) throws DaoException {
+        batchBaseDao.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean exists(StringIdKey key) throws DaoException {
+        return batchBaseDao.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public TaskTypeIndicator get(StringIdKey key) throws DaoException {
+        return batchBaseDao.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public List<StringIdKey> batchInsert(@SkipRecord List<TaskTypeIndicator> elements) throws DaoException {
+        return batchBaseDao.batchInsert(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchUpdate(@SkipRecord List<TaskTypeIndicator> elements) throws DaoException {
+        batchBaseDao.batchUpdate(elements);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", rollbackFor = Exception.class)
+    public void batchDelete(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        batchBaseDao.batchDelete(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean allExists(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        return batchBaseDao.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public boolean nonExists(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        return batchBaseDao.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<TaskTypeIndicator> batchGet(@SkipRecord List<StringIdKey> keys) throws DaoException {
+        return batchBaseDao.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<TaskTypeIndicator> lookup() throws DaoException {
+        return entireLookupDao.lookup();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @SkipRecord
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public List<TaskTypeIndicator> lookup(PagingInfo pagingInfo) throws DaoException {
+        return entireLookupDao.lookup(pagingInfo);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
+    public int lookupCount() throws DaoException {
+        return entireLookupDao.lookupCount();
+    }
+}