+++ familyhelper-project-sdk/src/main/java/com/dwarfeng/familyhelper/project/sdk/bean/entity/WebInputTaskTypeIndicator.java
@@ -1,0 +1,83 @@
+package com.dwarfeng.familyhelper.project.sdk.bean.entity;
+
+import com.alibaba.fastjson.annotation.JSONField;
+import com.dwarfeng.familyhelper.project.sdk.util.Constraints;
+import com.dwarfeng.familyhelper.project.stack.bean.entity.TaskTypeIndicator;
+import com.dwarfeng.subgrade.sdk.bean.key.WebInputStringIdKey;
+import com.dwarfeng.subgrade.stack.bean.Bean;
+import org.hibernate.validator.constraints.Length;
+
+import javax.validation.Valid;
+import javax.validation.constraints.NotEmpty;
+import javax.validation.constraints.NotNull;
+
+/**
+ * WebInput 任务类型指示器。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public class WebInputTaskTypeIndicator implements Bean {
+
+    private static final long serialVersionUID = -3068677962131260608L;
+
+    public static TaskTypeIndicator toStackBean(WebInputTaskTypeIndicator webInputTaskTypeIndicator) {
+        return new TaskTypeIndicator(
+                WebInputStringIdKey.toStackBean(webInputTaskTypeIndicator.getKey()),
+                webInputTaskTypeIndicator.getLabel(),
+                webInputTaskTypeIndicator.getRemark()
+        );
+    }
+
+    @JSONField(name = "key")
+    @Valid
+    private WebInputStringIdKey key;
+
+    @JSONField(name = "label")
+    @NotNull
+    @NotEmpty
+    @Length(max = Constraints.LENGTH_LABEL)
+    private String label;
+
+    @JSONField(name = "remark")
+    @NotNull
+    @NotEmpty
+    @Length(max = Constraints.LENGTH_REMARK)
+    private String remark;
+
+    public WebInputTaskTypeIndicator() {
+    }
+
+    public WebInputStringIdKey getKey() {
+        return key;
+    }
+
+    public void setKey(WebInputStringIdKey key) {
+        this.key = key;
+    }
+
+    public String getLabel() {
+        return label;
+    }
+
+    public void setLabel(String label) {
+        this.label = label;
+    }
+
+    public String getRemark() {
+        return remark;
+    }
+
+    public void setRemark(String remark) {
+        this.remark = remark;
+    }
+
+    @Override
+    public String toString() {
+        return "WebInputTaskTypeIndicator{" +
+                "key=" + key +
+                ", label='" + label + '\'' +
+                ", remark='" + remark + '\'' +
+                '}';
+    }
+}